utils: Introduce “counters” module.
The IDs returned by the Figshare API rely on enumerating numbers to
provide unique identifiers to records.  This is usually handled
by a relational database with an automatically incrementing counter.

We are using an RDF store instead, so we need to implement
concurrency-safe counters.  This module provides that.

This implementation implies only one instance of this program may
run at any time.

* src/rdbackup/utils/counters.py: New file.