backup: Also gather "totals" statistics.
There seems to be a difference in what is reported through the
breakdown statistics API calls, and the reported statistics on the
portal.  The portal uses a different API call.

I reported the issue to Figshare.  In the meantime, the "totals"
statistics we gather with the additions of this commit are the exact
numbers reported on the Figshare portal.

* sql/Database.sql: Add “ArticleTotals” and “CollectionTotals” tables.
* src/djehuty/backup/database.py: Implement ‘insert_totals_statistics’.
* src/djehuty/backup/figshare.py: Add the “totals” statistics.