api: Introduce a caching layer.
Our data doesn't change unless the graph name changes.  So executing
the same query multiple times always results in the same answer.  We
can store these results so that they can be retrieved more quickly
later on.  This commit implements a simple caching layer based on the
full SPARQL query sent to the SPARQL endpoint. Caching is only enabled
on a handful of calls that are most prone to poor performance.

* src/djehuty/api/cache.py: New file.
* src/djehuty/api/database.py: Use the cache layer for long-running queries.
* src/djehuty/api/ui.py: Set the cache directory.