api: html_templates: Move CSS bits for specific pages to specific pages.
Every page can have its own unique layout.  Before this commit, we
stored all styling in a single CSS file.  After this commit, we store
the page-specific styling in the <head> of the page.

This makes a small difference in favor of keeping the main stylesheet
short, and avoid affecting each individual's page with accidental
overlapping styles.

* src/djehuty/api/resources/html_templates/agriculture-animal-plant-sciences.html:
  Store CSS inline.
* src/djehuty/api/resources/html_templates/portal.html: Likewise.
* src/djehuty/api/resources/static/css/main.css: Remove page-specifics from
  the global stylesheet.