web: database: Implement session handling.
This commit introduces session tokens in the RDF store. Session tokens
are used as temporary identifiers with which a user can authenticate
in Djehuty.  A session token can be stored in a HTTP cookie, and the
procedures implemented in this commit can create and delete such
session tokens, and look up account information by a session token.

* src/djehuty/utils/counters.py: Add a counter for sessions.
* src/djehuty/web/database.py: Implement procedures around session tokens.
* src/djehuty/web/resources/sparql_templates/account_by_session_token.sparql: New file.
* src/djehuty/web/resources/sparql_templates/delete_session.sparql: New file.