web: wsgi: Remove statically defined tokens.
In commit ffddc3e0e912b8780f0a3f661a6d7464a46b0806 we added privileges
configuration to the configuration file, which includes the
“may-impersonate” property.

In this commit we remove the static tokens and use the privileges
configuration instead to implement the “may-impersonate” property.

So now we can define privileges for specific users in the
configuration file.  No such information is added to the database,
and no such information can be found in the source code.

* src/djehuty/web/wsgi.py: Remove static tokens, remove the Account class.