web: html_templates: my-data: Implement multi-file drag and drop.
This commit implements the bits to handle the case where a user drops
multiple files in the drag-and-drop container.  Doing a for-loop with
each file results in uploading all files at the same time.  Instead of
doing that, I went for uploading one-at-a-time by passing the logic
down to ‘perform_upload’.

* src/djehuty/web/resources/html_templates/depositor/my-data.html:
  Adjust styling.
* src/djehuty/web/resources/static/js/my-data.js: Implement multi-file
  drag and drop.