web: database: Implement “exclude_ids” parameter for ‘articles’.
When searching for articles to add to a collection, one wants to avoid
finding an article that is already in the collection.  A general way
to make sure these articles are excluded from the search is to pass a
list of article IDs to exclude.  That's what this commit implements.

* src/djehuty/web/database.py: Implement “exclude_ids” parameter for
  ‘articles’.
* src/djehuty/web/wsgi.py: Pass the “exclude_ids” parameter through in
  ‘api_private_articles_search’.