web: wsgi: Implement a Git server.
This commit implements a Git server by proxying to “git http-backend”.
It requires “git” to be available on the server-side.  Currently, only
the smart protocol is implemented. For the dumb protocol, one needs to
add direct file access methods.

The big missing piece is authentication.  The Git protocol allows for
HTTP basic authentication, which can be implemented separately from
the interaction with “git http-backend”.

* requirements.txt: Add “pygit2” dependency.
* src/djehuty/web/wsgi.py: Implement the Git server.