web: wsgi: Fix bug in behavior change in werkzeug 2.1.
Calling ‘get_json’ on a request that doesn't have a JSON Content-Type
will respond with a 400 error.  The previous behavior was to return
None in the ‘get_json’.

In the check for the “impersonate” parameter we rely on the old
behavior.  This commit checks the Content-Type before attempting to
call ‘get_json’.

* src/djehuty/web/wsgi.py: Add Content-Type check.