backup: Introduce the “is_editable” flag.
Before this commit, the latest version of a dataset was fetched from
the private API calls.  This is wrong.  In this commit, we treat the
version fetched from the private API call as unpublished.  For this
version, we set the “is_editable” flag.  The latest version is now
fetched using the public API call.

* sql/Database.sql: Implement “is_editable” flag.
* src/djehuty/backup/database.py: Likewise.
* src/djehuty/backup/figshare.py: Likewise.