web: database: Fix erroneous value for “files” and “bytes” in the overview.
The SPARQL aggregate functions don't remove duplicates. Using DISTINCT
would only de-duplicate files with the same size, which isn't what we
want to have here.

This commit makes the “repository_statistics_files” query return each
unique file as a row.  Then we sum the number of files and number of
bytes in Python instead of SPARQL.

* src/djehuty/web/database.py: Sum the results in Python.
* src/djehuty/web/resources/sparql_templates/repository_statistics_files.sparql:
  Let the query return a list of files and their sizes.