web: ui: Allow including configuration files.
This commit introduces the <include> configuration item that can be
used to include additional configuration files.  It works as
following:

  <?xml version="1.0" encoding="utf-8"?>
  <djehuty>
    <bind-address>127.0.0.1</bind-address>
    ...
    <include>djehuty-static-site.xml</include>
  </djehuty>

The additional configuration file must have the similar structure:

  <?xml version="1.0" encoding="utf-8"?>
  <djehuty>
    <static-pages>
      <page> ... </page>
    </static-pages>
  </djehuty>

The order in which items are parsed is unspecified, so in the cases
where configuration items are duplicated there is no guarantee one
will always overwrite the other.

* src/djehuty/web/ui.py: Separate configuration file parsing into
  ‘read_configuration_file’, and use it to recurse into “include”
  files.