backup: Do not use a relational database as intermediate step.
This commit overhauls the backup functionality, breaking both the
command-line interface as well as the data model.

The backup tool serializes RDF triplets to standard output, which
means it is wise to redirect standard output to a file. Here's an
example:
  djehuty backup --token=... 2> errors.log | gzip --stdout > triplets.n3.gz

The resulting RDF no longer uses link tables, but instead uses
rdf:List structures.  Follow-up commits will address the implications
of this in the “web” part of Djehuty.

* requirements.txt: Remove the “mysql-connector-python” dependency.
* rpmbuild/SPECS/djehuty.spec.in: Likewise.
* src/djehuty/backup/database.py: Generate RDF output instead of
  inserting into a MySQL database.
* src/djehuty/backup/figshare.py: Adapt to changes in “database.py”;
  Add ability in ‘get_institutional_accounts’ to return a single
  account record; Avoid duplicated requests for author details.
* src/djehuty/backup/ui.py: Adapt to changes in “database.py” and
  “figshare.py”; Remove MySQL database parameters.
* src/djehuty/ui.py: Change command-line interface for “backup”.
* src/djehuty/utils/rdf.py: Simplify the namespace for row types.