backup: database: Implement ArticleContainers.
Article IDs are shared among article versions, published or
unpublished.  Up until now we've annotated special versions using
the “is_editable” and “is_latest” flags.  This commit introduces
a new structure that takes a step back and describes the situation
slightly better.

An ArticleContainer can have a single “draft” version, which is
unpublished. This is the same version that was marked as
“is_editable”.

Published versions are an rdf:List stored under the predicate
“col:published_versions”.  The latest published version is the last
element of this list.  It can be retrieved easily by using the rdf:nil
property of the last list item.

* src/djehuty/backup/database.py: Implement ArticleContainers.