web: wsgi: Adapt ‘api_article_versions’ to new data model.
As the first procedure to be adapted, let's scrutinize the details.
The “article_id” parameter is passed along as string, regardless of
whether it contains the Figshare ID (integer) or something else.

With ‘parses_to_int’ we distinguish between referring to an article by
its Figshare ID or its UUID.  Doing so enables us to pass either
“container_uri” or “article_id” to ‘dataset_versions’.

So after this commit, both of these URLs would produce the same results:
  /v2/articles/306ee4ac-6bb8-4647-9f19-3c80b22df020/versions
  /v2/articles/12763700/versions

... given that that UUID is the container URI for the article with
Figshare ID 12763700.

* src/djehuty/web/wsgi.py: Adapt ‘api_article_versions’ to new data model.