web: database: Implement ‘delete_associations’.
This commit implements a procedure to remove a triplet of which its
object is an rdf:List.  It is essential for update operations in the
new state-graph data model.

It is purposefully limited to delete elements from draft items in a
container, so it cannot be used to delete data from published items.

* src/djehuty/web/database.py: Implement ‘delete_associations’.
* src/djehuty/web/resources/sparql_templates/delete_associations.sparql:
  New file.