web: database: Implement ‘update_author_list’.
This procedure implements a method to replace an author list in a
draft article.  Our rdf:Lists have an index column per list item, and
that makes it hard to modify the rdf:Lists in-place.  A simpler
strategy is to rebuild the list instead.  The run-time cost is that
one needs to always read the entire list, modify it in-memory, and
write the entire new list.

The ‘update_author_list’ is the first procedure to be implemented
using this new strategy.

* src/djehuty/web/database.py: Implement ‘update_author_list’.