Replace “xml.etree” with “defusedxml”.
The documentation of “xml.etree.ElementTree” [1] states that it's not
safe against user input.

The XML parser from this module is used only to take input from
trusted sources:
1. Reading the configuration file upon program start.
2. Reading XML from the OPeNDAP server maintained by our team.

The alternative library called “defusedxml” [2] provides a drop-in
replacement for “xml.etree” that is recommended for parsing
potentially malicious user input.

This patch switches to “defusedxml” to prevent creating a potential
security risk when growing the codebase in the future at the cost
of adding “defusedxml” as dependency.

[1] https://docs.python.org/3/library/xml.etree.elementtree.html
[2] https://docs.python.org/3/library/xml.html#the-defusedxml-package

* guix.scm: Add the “defusedxml” dependency.
* requirements.txt: Likewise.
* src/djehuty/backup/database.py: Use “defusedxml” in favor of “xml.etree”.
* src/djehuty/web/ui.py: Likewise.