web: Use a separate UUID for Git repositories.
The Git repository URL was bound to the container UUID of a dataset.
This introduces two problems:
1. The Git URL is guessable for published articles;
2. The Git repository is shared amongst dataset versions.

This commit solves both problems by generating a unique UUID for Git
repositories per version of a dataset.

The consequence is that a user needs to update the Git URL after
publishing a dataset.

* src/djehuty/web/database.py: Add “git_uuid” property.
* src/djehuty/web/resources/html_templates/depositor/edit-dataset.html:
  Use the “git_uuid” property for the repository URL.
* src/djehuty/web/resources/sparql_templates/datasets.sparql: Add the
  “git_uuid” property as output variable.
* src/djehuty/web/wsgi.py: Use “git_uuid” instead of “dataset_id” for
  the Git interaction.