web: ui: Exit when a configuration file cannot be found.
Before this commit, when <include>ing a non-existent file, one would
receive a file-not-found message for both the non-existent file as
well as the originating file before exiting the program.

This commit changes the implementation to exit early so that one only
sees the file-not-found message for the actual missing file.

* src/djehuty/web/ui.py: Raise a “SystemExit”, effectively terminating
  the program.