web: validator: Implement validation error stacking.
Before this commit, when using the validator procedures, the first
error would raise an exception, and that limited the reporting of
errors to only reporting the first one.

This commit adds an “error_list” parameter to the validator procedures
which will be appended to when an error occurs.  With this in place,
reporting multiple errors is possible by passing an empty list as
“error_list” and checking at a later point whether the “error_list” is
empty.

* src/djehuty/web/validator.py: Implement “error_list” and its behavior.
* src/djehuty/web/wsgi.py: Adapt to modified interface of ‘order_direction’.