web: wsgi: Make updating file lists concurrency-safe.
In the time between getting the list of files and writing the new list
of files to the database another file could've been uploaded, leading
to a loss of information.  To prevent that, we ensure only one such
operation can be done at a single time by using a thread lock or uwsgi
lock.

* src/djehuty/web/wsgi.py: Implement thread lock and uwsgi lock when
  updating file lists.