web: sparql_templates: accounts: Bring back numeric account ID.
In an effort to run the backup tool on a Djehuty instance rather than
Figshare, I noticed that the lack of numeric account IDs can be hard
for scripts that assume “id” to be set.  So in the name of
backward-compatibility, this commit brings back numeric account IDs.

This commit only implements it for migrated accounts. Newly created
accounts won't have a numeric ID.

* src/djehuty/web/formatter.py: Bring back the numeric account ID.
* src/djehuty/web/resources/sparql_templates/accounts.sparql: Likewise.