web: sparql_templates: publish_draft_dataset: Work around Virtuoso bugs.
Bug 1:
Apparently, rdf:nil as object equals an unbound value.  The
work-around consists of string-matching the value for rdf:nil.  This
should be cross-endpoint compatible, but comes at a (small)
performance cost.

Bug 2:
Querying an inserted literal with an explicit xsd:string type by
explicitly matching the type yields no results.  The work-around here
is to not explicitely define the xsd:string type at insertion-time.
This should be cross-endpoint compatible because in section 17.4.2.7
of the SPARQL 1.1 standard it is stated:
  "If the literal is a simple literal, return xsd:string".

Thus, I assume an literal without a specified type (a simple literal)
will be interpreted as a typed literal when needed where its type is
an xsd:string.

* src/djehuty/web/resources/sparql_templates/publish_draft_dataset.sparql:
  Work around Virtuoso bug regarding rdf:nil; The
  djht:published_datasets is an unversioned list;