web: expose versionless doi in landing pages when relevant
The latest version of a dataset/collection can be accessed with a doi
with or without version number. Both dois are displayed and the difference
is explained in the UI when hovering over the doi.

Also improve handling of containers for datasets and collections
and tweak css to better match other pages (use of rounded corners)

* src/djehuty/web/wsgi.py: adapt ui_dataset() and ui_collection()
* src/djehuty/web/resources/html_templates/public_metadata.html:
  add versionless doi + info
* src/djehuty/web/resources/static/pub.css: add style for versionless doi
  and round some corners