web: database: Improve portability of interpreting typed literals.
Blazegraph doesn't distinguish “typed-literal” from “literal”, and
without datatype, it is assumed to be an xsd:string.  Instead of
testing for “typed-literal” we directly test for the availability of
“datatype”, which makes it more portable.

Furthermore, Blazegraph returns “false” and “true” for boolean values,
while Virtuoso returns 0 and 1.  This commit implements the extra
logic for interpreting “false” and “true”.

* src/djehuty/web/database.py: Improve portability of interpreting
  typed literals.