web: database: Prepare sessions for two-factor authentication.
This commit continues the work started in edfbf36 to add 2FA for
privileged users.  It adds a “djht:active” property to all sessions
which invalidates all sessions created before this commit.  For
accounts that require 2FA, ‘insert_session’ adds a “djht:mfa_token”
property and returns its value.

The ‘sessions’ procedure can now be passed an “mfa_token”.  When done
so it looks for an inactive session for “account_uuid” with
“session_uuid” and “mfa_token”.

A follow-up commit implements the user interface bits.

* src/djehuty/web/database.py: Prepare session handling for 2FA.
* src/djehuty/web/resources/sparql_templates/account_by_session_token.sparql: Likewise.
* src/djehuty/web/resources/sparql_templates/account_sessions.sparql: Likewise.
* src/djehuty/web/resources/sparql_templates/update_session.sparql: Likewise.
* src/djehuty/web/wsgi.py: Adjust the interface change of
  ‘insert_session’ which returns three values now.