web: wsgi: Implement two-factor authentication.
Accounts with elevated privileges require 2FA upon login.  This commit
implements builds on fb507a9 to implement the logic for the user
interface.  The summary is (1) user logs in using an identity
provider, (2) when 2FA is required, the user gets an inactive token
and gets redirected to the 2FA token page (3) after entering the 2FA
token the session is activated.

* src/Makefile.am: Register new files.
* src/djehuty/web/resources/html_templates/activate_2fa_session.html: New file.
* src/djehuty/web/resources/html_templates/email/2fa_token.html: New file.
* src/djehuty/web/resources/html_templates/email/2fa_token.txt: New file.
* src/djehuty/web/wsgi.py: Extend ‘ui_login’ and implement ‘ui_activate_session’.