web: Implement creating private links for published datasets.
This commit is quite involved, so prepare. :)
Up until this point, we assumed that private links were only generated
for unpublished (draft) datasets.  The general procedures
‘update_item_list’ and ‘delete_associations’ used to have a built-in
protection to prevent modifying published items by getting an item by
its container-[draft]->item link.

This commit overhauls that.  Instead of passing the container UUID to
the above mentioned procedures, one now passes the item UUID
instead. The container is found by the inverse link
item-[container]->container.

This change affects the interface of various other procedures such as
‘update_dataset’ and ‘update_collection’.  Therefore the commit
affects a large surface of the codebase.  We have to apply this change
in a single commit to keep things consistent.

* src/djehuty/web/database.py: Use “item_uuid” instead of
  “container_uuid” in ‘update_item_list’ and ‘delete_associations’;
  Apply interface change throughout.
* src/djehuty/web/wsgi.py: Use dataset UUID instead of container UUID
  for private links pages; Adapt to interface changes of
  ‘update_item_list’ and ‘delete_associations’.
* src/djehuty/web/resources/html_templates/depositor/dataset-private-links.html:
  Use dataset UUID instead of container UUID.
* src/djehuty/web/resources/html_templates/depositor/my-data.html: Add
  private links button to published datasets.
* src/djehuty/web/resources/sparql_templates/container_items.sparql:
  Implement reverse item <-> container lookup.
* src/djehuty/web/resources/sparql_templates/datasets.sparql: Likewise.
* src/djehuty/web/resources/sparql_templates/delete_associations.sparql:
  Likewise.
* src/djehuty/web/resources/sparql_templates/update_dataset.sparql:
  Adapt to interface change in ‘update_dataset’.