web: Ensure only latest versions of datasets are added to collections pt.2
In commit d13da58cb75a8df3c530e244a878515e19133d46 it was ensured that
only latest versions of datasets were shown as a search result for adding
them to a collection. The result list of added datasets still showed
multiple versions. This is fixed in the current commit.

* src/djehuty/web/wsgi.py: adapt api_private_collection_datasets()