web: wsgi: Adapt “collect” button to new collection creation strategy.
The “collect” button on a dataset landing page was built under the
assumption that there is always a draft version of a collection.
However, we switched to a strategy in which a user must explicitely
create a new version of a published collection before a draft version
is available.

This commit implements the bits to automatically create a draft
collection from its published equivalent, and adapts the query to list
the collections of an account to list both draft and latest published
versions.

* src/djehuty/web/resources/sparql_templates/collections_by_account.sparql:
  Adapt query to list drafts and latest published versions of a collection.
* src/djehuty/web/wsgi.py: Implement the bits to automatically create
  a draft collection.