web: wsgi: Make ‘ui_download_file’ more precise.
This commit fixes a problem with downloading version-specific files.
When version 2 of a dataset contains a file that version 1 doesn't
contain, the download would fail because the look up of the dataset
ended up with the oldest version.  This commit first looks up the file
metadata and then looks up the dataset that belongs to it.

An extra check has been added to ensure the file metadata and the
dataset metadata originate from the same container.

* src/djehuty/web/resources/sparql_templates/dataset_files.sparql:
  Always link to the container, not only when a dataset_uri is present.
* src/djehuty/web/wsgi.py: Make ‘ui_download_file’ more precise.