web: wsgi: Implement streaming multipart/form-data.
By default, werkzeug writes large files to a temporary location before
allowing us to access the file contents to store it elsewhere, and to
calculate a checksum.

This commit implements a streaming variant of multipart/form-data, in
which it stores the contents in the final destination, and it computes
the MD5 while writing the file contents.

After an upload is complete, the MD5 of the received contents will be
immediately available.  This commit assumes that what is written to
disk is the same as what was passed to ‘write’ (intended to be stored
on disk).

* src/djehuty/web/wsgi.py: Implement streaming multipart/form-data.