web: sparql_templates: update_dataset: Allow erasing the resource title/DOI.
Because empty values are processed to None, we don't differentiate
between value-was-not-provided and value-was-provided-but-empty.  Not
necessarily a problem, unless we allow partial updates without
removing other values.

A common case of removing the resource DOI and resource title is
accomodated by this experiment to always remove these values,
regardless of whether a new value is provided.

* src/djehuty/web/resources/sparql_templates/update_dataset.sparql:
  Always remove the previous values of “djht:resource_title” and
  “djht:resource_doi”.