web: wsgi: Restructure and revise ‘ui_download_file’.
This commit changes the logic of the ‘ui_download_file’ procedure to
hopefully make it more comprehensible.  In addition it improves the
logging of why access was granted (for auditing purposes) and it
solves a bug where a cached “is_embargoed” would not give access to a
lifted embargoed file.

* src/djehuty/web/wsgi.py: Restructure ‘ui_download_file’ and improve
  its logging.