web: wsgi: Fix inaccessible embargoed datasets by owners and reviewers.
The owner of dataset should be able to bypass the restriction of embargo.
This is done by passing owner's 'account_uuid' to `__file[s]_by_id_or_uri()`.
For draft datasets, the parameter is passed to the function. However,
for published datasets, passing the parameter to the function was
missing.

* src/djehuty/web/wsgi.py: Pass 'account_uuid' to `__file[s]_by_id_or_uri()`