web: Improvement for incomplete upload and its MD5 checksum
This commit is (1) to create MD5 checksum even when upload is incomplete and
(2) to display a warning text next to filename that is not completely uploaded.
To implement this, "is_incomplete" was added to the SPARQL templates for files.
In addition, missing MD5 checksum for zero-byte files is also fixed by this commit.

* src/djehuty/web/database.py: Pass "is_incomplete" to the SPARQL
  templates for files.
* src/djehuty/web/formatter.py: Add "is_incomplete" to return it in JSON.
* src/djehuty/web/resources/html_templates/depositor/edit-dataset.html:
  Add a style of ".file-incomplete-warning".
* src/djehuty/web/resources/sparql_templates/dataset_files.sparql: Add
  "is_incomplete".
* src/djehuty/web/resources/sparql_templates/update_file.sparql: Add
  "is_incomplete".
* src/djehuty/web/resources/static/js/edit-dataset.js: Display the warning text.
* src/djehuty/web/wsgi.py: Add exceptions for incomplete upload.