web: Implement automatic-login for development setups.
Djehuty relies on an external identity provider to handle logging in.
This is a bit tedious for development setups.  So this commit
implements a “automatic-login-email” configuration option to
automatically log in the account with that particular e-mail address
without needing a third-party identity provider.

It only works when “production” is set to 0, and when no other
identity provider is configured.

* etc/djehuty/djehuty-example-config.xml: Add example for
  “automatic-login-email”.
* src/djehuty/web/resources/html_templates/layout.html: Add bits for
  the “automatic-login” identity provider.
* src/djehuty/web/ui.py: Implement and use
  ‘read_automatic_login_configuration’.
* src/djehuty/web/wsgi.py: Implement the automatic login action.