web: wsgi: Add from_draft parameter in __update_item_doi.
When reserving or updating a DOI has been failed while publishing a dataset,
there is no way to reuse __update_item_doi() to fix the DOI because the
function tries to get the dataset's metadata from its draft instead of its
specified version. This patch allows to get metadata from draft or the
specified version.

* src/djehuty/web/wsgi.py: Add from_draft parameter in __update_item_doi.