utils: Remove unused module.
The idea to render the Markdown in the descriptions gave rise to this
module to convert the existing HTML subset that is allowed in the
description field.  If we pursue this idea, this module could be used
for a one-time conversion of existing HTML.  So feel free to resurrect
this code when there is a use for it.

* src/Makefile.am: Unregister the “html2md” module.
* src/djehuty/utils/html2md.py: Remove module.