web: database: Allow re-using an account record in ‘__may_execute_role’.
When rendering an HTML template, a couple of parameters are passed
along that invoke ‘may_*’ which in turn invokes ‘__may_execute_role’.
This repeatedly queries the SPARQL endpoint for the same information.

This commit adds a parameter to ‘__may_execute_role’ so that the
repeated queries to the SPARQL endpoint can be reduced to only a
single query.

* src/djehuty/web/database.py: Add “account” parameter to
  ‘__may_execute_role’ and pass it through the ‘may_*’ procedures.
* src/djehuty/web/wsgi.py: Pass along the “account” record to the
  ‘may_*’ invocations.