web: wsgi: Clean up error-catching of various API calls.
This commit simplifies a pattern that was evolved from the
use of ‘__dataset_by_id_or_uri’;  We used to need to catch IndexError
because we would directly take the first list item of a call to
‘db.datasets’.  However, ‘__dataset_by_id_or_uri’ catches the
IndexError, and returns None.  So to catch a missing/inaccessible
dataset, we need to check for None, rather than catch IndexError.

* src/djehuty/web/wsgi.py: Simplify error-checking in a couple of
  cases and clean up.