web: Allow removing all categories.
When removing already-set categories, the deletion would be omitted
due to the falsiness of an empty list in Python.  This was by design,
to never accidentally lose the categories associated with a dataset.

However, this made removing all categories impossible.  This commit
addresses this by differentiating between None and an (non-)empty
list.

* src/djehuty/web/database.py: Differentiate between instances of
  lists and None in ‘update_dataset’.
* src/djehuty/web/wsgi.py: Explicitly pass None as categories
  parameter.