web: database: Improve error handling of DB response (400 Bad Request)
This commit prevents an infinite loop caused by a syntax error query.

RDFLib has a mechanism that keeps failed transactions in the internal queue
to re-execute them when the next update query comes in. The problem is
that this re-execution occurs even if the failed transactions have a syntax
error.

Therefore, if the cause of the failure is due to a syntax error, the
transaction should be rolled back.

This commit also sets autocommit = False in sparqlstore's init because commit()
is explicitly called right after self.sparql.update().

* src/djehuty/web/database.py: Add a rollback in case server response is
  400.