web: Add a new 'Recent System Log' report
This commit adds a new report, 'Recent System Log'. The purpose of this
report is to check if the service has been restarted recently without sshing
to the server.

The log file is retrived from the end up to 'limit_log_size'. The default
limit size is 10 MB. The report shows only log messages generated by
'djehuty.ui' logger.

* src/djehuty/web/resources/html_templates/admin/reports/dashboard.html:
  Add a new link pointing to the report.
* src/djehuty/web/resources/html_templates/admin/reports/recent_system_log.html:
  A new HTML template of the report.
* src/djehuty/web/wsgi.py: Implement 'ui_admin_reports_recent_system_log'.