web: wsgi: Preserve 'embargo_until_date' while reserving/saving a DOI.
This commit is a follow-up from be8d96c. When a dataset is
published in production, the value of 'embargo_until_date' is deleted
while reserving/saving a DOI.

As the value of 'embargo_until_date' is unconditionally deleted in the
SPARQL query template, the value is explicitly passed to the SPARQL
query template by this commit.

* src/djehuty/web/wsgi.py: Pass the value of 'embargo_until_date' when
  self.db.update_dataset() is called.