web: html_template: Prevent loading thumbnails from an unavailable external source.
There are datasets that have thumbnails stored in external sources.
Among the external sources, there is one source that returns 403
Forbidden response to all thumbnail requests. This results in displaying
'alt' messages instead of a default thumbnail in tile view.

This commit fixes it so that thumbnails from the external source are replaced
with the default SVG thumbnail. This fix can be removed once the
external source has been replaced in the database.

* src/djehuty/web/resources/html_templates/search.html: Don't load the
  thumbnails from their sources if their sources start with
  'https://ndownloader'.