web: wsgi: Avoid outputting permissions in datasets query.
This commit attempts to solve a bug that originates from putting the
contributor permissions in the output variables of the datasets query
and the change in 1e73d5a which changed the row output behavior of
datasets with multiple contributors.

It solved it by decoupling the permissions from the dataset metadata.
Because of this, a more elaborate mechanism to check whether
contributor permissions are needed, has to be introduced
(see ‘__needs_collaborative_permissions’).

* src/Makefile.am: Register new file.
* src/djehuty/web/database.py: Implement
  ‘item_collaborative_permissions’.
* src/djehuty/web/resources/html_templates/depositor/edit-dataset.html:
  Simplify.
* src/djehuty/web/resources/sparql_templates/dataset_files.sparql:
  Remove collaborator permissions output variables.
* src/djehuty/web/resources/sparql_templates/datasets.sparql: Likewise.
* src/djehuty/web/resources/sparql_templates/item_collaborative_permissions.sparql:
  New file.
* src/djehuty/web/wsgi.py: Implement and use
  ‘__needs_collaborative_permissions’.