web: wsgi: Implement ‘api_v3_dataset_git_languages’.
This API call attempts to close the gap in API expected from a Git
platform in addition to the Git protocol.  It implements a simple
method of counting lines per programming language in a Git repository,
and it provides a new API endpoint to communicate that information.

It skips binary files and files larger than 5MB because we don't
expect code to be in those forms.

* src/djehuty/utils/constants.py: Implement “filetypes_by_extension”.
* src/djehuty/web/wsgi.py: Implement ‘__git_files_by_type’ and
  ‘api_v3_dataset_git_languages’.