web: sparql_templates: datasets: Prevent locking out owner as collaborator.
In the erroneous case that a dataset owner is also a collaborator, the
permissions of the collaboration are used, which can lead to the owner
locking him/herself out editing the dataset.  This commit addresses it
by adding an explicit check that determines whether the dataset
“is_shared_with_me”, which in turn prevents from the restrictions from
being applied for the owner.

* src/djehuty/web/resources/sparql_templates/datasets.sparql: Add
  check for whether the “collaborator_account” is the owner.