utils: convenience: Implement 'split_string' can check use of quotes.
This commit deprecates 'split_delimitered_string' and introduces a new
function, 'split_string'. This new function provides an option to ignore
delimiters when they are part of word(s) surrounded by single/double quotes.
Another option is `maxsplit`, the same as str.split()'s

This commit fixes a no-keyword-search-result bug occuring when a keyword
is clicked on a published dataset page.

* src/djehuty/utils/convenience.py: Replace 'split_delimitered_string'
  with the new function, 'split_string'.
* src/djehuty/web/database.py: Apply 'split_string' to dataset's
  search filters.
* src/djehuty/web/wsgi.py: Apply 'split_string'.