web: wsgi: Adapt `is_featured` logic for `api_v3_groups`.
This commit adapts a new parameter `is_featured` for `api_v3_groups`.
Therefore, `djht:is_featured` predicate is added to `djht:InstitutionGroup`.
This new predicate can be used to display only featured institutions in
the main portal page. Another example is to use for the institution list
in the search filter.

* src/djehuty/utils/rdf.py: Fix a typo in a comment.
* src/djehuty/web/wsgi.py: Use `is_featured` in /v3/groups API.
* src/djehuty/web/resources/sparql_templates/group.sparql: Apply the new
  predicate.
* src/djehuty/backup/resources/groups.json: Add `is_featured` property
  to 4TU members.
* src/djehuty/web/database.py: Add `is_featured` parameter and apply it
  to sparql_filter().