web: wsgi: Implement “storage” configuration option.
This commit implements a new way to add multiple storage locations
which will each be checked to see if a file exists on the given
storage location.  The configuration looks like this:

<djehuty>
  ..
  <storage>
    <location>/data-1</location>
    <location>/data-2</location>
    <location quirks="1">/data-3</location>
  </storage>
  ...
</djehuty>

Eventually, this mechanism should replace the “primary-storage” and
“secondary-storage” configuration options, but for the moment, either
configuration is allowed.

The benefit of this method is that an arbitrary number of storage
locations can be specified, allowing one to move files between storage
back-ends and having it transparently handled by Djehuty.

* src/djehuty/web/database.py: Add “storage_locations” variable.
* src/djehuty/web/resources/sparql_templates/dataset_files.sparql:
  Output filename without its full path.
* src/djehuty/web/ui.py: Read the storage configuration.
* src/djehuty/web/wsgi.py: Use the new storage configuration to find
  the filesystem location of a file.