web: formatter: Add `is_featured` property to `format_group_record`.
In abbb3bc commit, `is_featured` predicate was added to `api_v3_group`.
That commit allows to filter groups by `is_featured`. However, the key
name and its value of `is_featured` are not included in the JSON result
when the API is called.

This commit implements that the key name and its value are included to
the result when the URL /v3/group is called.

* src/djehuty/web/formatter.py: Add `is_featured` property to the format
  for groups.