web: search: Make institutions in sidebar dynamic.
This commit removes the hard-coded institution list in sidebar. Instead,
generates the list dinamically.

Furthermore, institutions ending with " Students" become hidden in the
list. If the selected institutions have separate IDs for students, the
group IDs for students are also queried together to the search API
on the Javascript side.

* src/djehuty/web/wsgi.py: Pass the result of db.group() to the Jinja
  template for search; Allow multiple group IDs in the v3 search API.
* src/djehuty/web/resources/html_templates/search.html: Render
  institutions received from wsgi.py
* src/djehuty/web/resources/static/js/search.js: Apply 'Show More' button
  for featured institutions; If the selected institutions have separate
  groups for students, add the groups too to the search API call.