web: wsgi: Handle missing files in ‘ui_download_all_files’.
Downloading a ZIP of all files may fail due to an inaccessible file on
the filesystem.  This would create a Traceback in the logs.  This
commit reports the missing file as an error, but skips it for
inclusion in the ZIP, making the download work for all files except
the missing ones.

* src/djehuty/web/wsgi.py: Skip unreachable files in
  ‘ui_download_all_files’.