web: wsgi: Implement v2 API endpoint to publish a dataset.
In a review of API endpoint of the v2 API, we found that the publish
endpoint is missing (which was deliberate).  But since the v3 endpoint
is compatible with the specification for the v2 endpoint, we can add
the endpoint to check off the box for this v2 API endpoint.

* src/djehuty/web/wsgi.py: Add endpoint for
  “/v2/account/articles/<dataset_id>/publish”.